# SubTrack/packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SubTrack API
  version: 1.0.0
  description: Contract metadata and approved operational endpoints.
paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: Health check
      responses:
        '200':
          description: API process is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        default:
          $ref: '#/components/responses/Problem'
  /readyz:
    get:
      operationId: getReadyz
      summary: Readiness check
      responses:
        '200':
          description: API is ready to receive requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessResponse'
        default:
          $ref: '#/components/responses/Problem'
  /v1/privacy/settings:
    get:
      operationId: getPrivacySettings
      summary: List open_book consent settings for all households (AC1)
      parameters:
        - in: header
          name: X-Caller-Id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of consent settings, one entry per household.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentSetting'
        '400':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/privacy/open-book:
    patch:
      operationId: setOpenBook
      summary: Toggle open_book for a household scope (AC1, AC4)
      description: >
        Sets the caller's open_book consent for the specified household.
        Default is OFF. Only the caller themselves can change their own setting —
        admins cannot alter other members' privacy (AC4).
        Updates take effect within 5 seconds (cache TTL, AC1).
        Writes a hash-chained audit event.
      parameters:
        - in: header
          name: X-Caller-Id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetOpenBookBody'
      responses:
        '200':
          description: Updated consent setting.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentSetting'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/privacy/preview-as/{householdId}:
    get:
      operationId: previewAsHousehold
      summary: Preview subscriptions as a household member (AC3)
      description: >
        Evaluates subscription visibility using the exact policy from ST-048.
        Returns safe subscription summaries — no raw transactions, balances,
        account numbers, or PII. Always-private subscriptions are hidden from
        non-owners even when the owner has open_book=ON (AC2).
      parameters:
        - in: path
          name: householdId
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Caller-Id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Visible subscription summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionSummary'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/data-rights/export:
    post:
      operationId: requestDataExport
      summary: Request a full personal data export (AC1)
      description: >
        Assembles a ZIP archive containing export.json (all personal data) and
        subscriptions.csv. Returns a single-use download token valid for 24 hours.
        Scoped only to the caller's own data.
      parameters:
        - in: header
          name: X-Caller-Id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: Export job accepted; archive is ready at the returned URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/data-rights/export/{token}:
    get:
      operationId: downloadDataExport
      summary: Download a previously requested data export (AC1, AC3)
      description: >
        Returns the ZIP file for a valid, non-expired export token.
        Returns 404 for unknown or expired tokens — the response does not
        distinguish between the two cases (AC3).
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ZIP archive download.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/data-rights/account:
    delete:
      operationId: deleteAccount
      summary: Permanently delete account and personal data (AC2)
      description: >
        Requires BankID or OTP re-authentication via `reAuthToken`.
        If the caller is the sole admin of a household with remaining members,
        returns 400 requiring role transfer first.
        Nullifies PII fields (email, phone, externalId) and sets deletedAt.
        Audit log entries are kept with pseudonymous identifiers only (AC2).
      parameters:
        - in: header
          name: X-Caller-Id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountBody'
      responses:
        '204':
          description: Account and personal data deleted.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'

  /v1/version:
    get:
      operationId: getVersion
      summary: API version
      responses:
        '200':
          description: API semantic version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
        default:
          $ref: '#/components/responses/Problem'
components:
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: ok
    ReadinessResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: ready
    VersionResponse:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: string
          description: Current API semantic version.
    SetOpenBookBody:
      type: object
      additionalProperties: false
      required: [householdId, openBook]
      properties:
        householdId:
          type: string
          format: uuid
        openBook:
          type: boolean
          description: true = share subscriptions with household peers; false (default) = private.

    ConsentSetting:
      type: object
      additionalProperties: false
      required: [householdId, openBook, updatedAt]
      properties:
        householdId:
          type: string
          format: uuid
        openBook:
          type: boolean
        updatedAt:
          type: string
          format: date-time

    SubscriptionSummary:
      type: object
      additionalProperties: false
      required: [id, categoryCode, cadence, status]
      properties:
        id:
          type: string
          format: uuid
        customName:
          type: string
          nullable: true
        categoryCode:
          type: string
        cadence:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, SEMIANNUAL, ANNUAL, CUSTOM]
        status:
          type: string
          enum: [DETECTED, TRIAL, ACTIVE, PAUSED, CANCELLED, ARCHIVED, REJECTED]
        alwaysPrivate:
          type: boolean
          description: Only present for the subscription owner — never exposed to other members.

    ExportJobResponse:
      type: object
      additionalProperties: false
      required: [token, downloadUrl, expiresAt]
      properties:
        token:
          type: string
          description: Single-use download token (64 hex chars).
        downloadUrl:
          type: string
          description: Relative URL to download the ZIP archive.
        expiresAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp after which the download token expires.

    DeleteAccountBody:
      type: object
      additionalProperties: false
      required: [reAuthToken]
      properties:
        reAuthToken:
          type: string
          description: >
            BankID or OTP re-authentication credential. Required before personal
            data erasure. With A5 Assumed (dev/simulator), any non-empty string
            is accepted.

    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required: [type, title, status]
  responses:
    Problem:
      description: An error response following RFC 9457.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'